@@ -1,11 +1,6 @@
-import {
-  client,
-  customersService,
-  customerGroupService
-} from './services'
-var nconf = require('nconf')
-
-var customersData = require(process.cwd() + '/data/customers.json')
+import { client, customersService, customerGroupService } from './services'
+const nconf = require('nconf');
+const customersData = require(process.cwd() + '/data/customers.json')
 
 const addCustomer = (cData) => {
   const updateRequest = {
@@ -16,9 +11,9 @@ const addCustomer = (cData) => {
 
   client.execute(updateRequest)
     .catch(error => console.log(error.body.errors))
-}
+};
 
-const importCustomerGroups = () => {
+const importCustomerGroups = () => {  
   var customerGroups = require(process.cwd() + '/data/customer-groups.json')
   customerGroups.forEach(element => {
     const updateRequest = {
@@ -30,9 +25,16 @@ const importCustomerGroups = () => {
     client.execute(updateRequest)
       .catch(error => console.log(error.body.errors))
   })
-}
+};
 
-const deleteAllCustomers = () => {
+const deleteAllCustomers = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: customersService
       .build(),
@@ -55,12 +57,15 @@ const deleteAllCustomers = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All customers are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All customers are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
-}
+};
 
-const deleteCustomerGroups = () => {
+const deleteCustomerGroups = (cb) => {
   const request = {
     uri: customerGroupService
       .build(),
@@ -83,17 +88,25 @@ const deleteCustomerGroups = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All customer groups are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All customer groups are succesfully deleted');
+    return cb()
   })
     .catch(err => console.log(err))
 }
+
 
 if (nconf.get('clean')) {
   deleteAllCustomers()
   deleteCustomerGroups()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing customers...');
   importCustomerGroups()
   customersData.forEach(element => {
     addCustomer(element)
   })
+}
+
+module.exports = {
+  deleteAllCustomers,
+  deleteCustomerGroups
 }
