@@ -1,9 +1,6 @@
-import {
-  client,
-  projectService
-} from './services'
-
-var projectData = require(process.cwd() + '/data/project.json')
+import { client, projectService } from './services'
+const nconf = require('nconf');
+const projectData = require(process.cwd() + '/data/project.json')
 
 const importProjectData = () => {
   const request = {
@@ -54,10 +51,10 @@ const importProjectData = () => {
         })
         .catch(err => console.log(err))
     })
-    .then(() => {
-      console.log('\x1b[32m%s\x1b[0m', 'Project data is succesfully imported')
-    })
     .catch(err => console.log(err))
 }
 
-importProjectData()
+if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing project data...');
+  importProjectData()
+}
