@@ -1,15 +1,11 @@
-import {
-  client,
-  typesService
-} from './services'
-
-var nconf = require('nconf')
+import { client, typesService } from './services'
+const nconf = require('nconf')
 
 const importTypes = () => {
-  var channelTypes = require(process.cwd() + '/data/channel-types.json')
-  var customerTypes = require(process.cwd() + '/data/customer-types.json')
-  var orderTypes = require(process.cwd() + '/data/order-types.json')
-  var allTypes = [channelTypes, customerTypes, orderTypes]
+  const channelTypes = require(process.cwd() + '/data/channel-types.json')
+  const customerTypes = require(process.cwd() + '/data/customer-types.json')
+  const orderTypes = require(process.cwd() + '/data/order-types.json')
+  const allTypes = [channelTypes, customerTypes, orderTypes]
 
   allTypes.forEach(type => {
     type.forEach(element => {
@@ -22,9 +18,16 @@ const importTypes = () => {
         .catch(error => console.log(error.body.errors))
     })
   })
-}
+};
 
-const deleteTypes = () => {
+const deleteTypes = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: typesService
       .build(),
@@ -47,13 +50,20 @@ const deleteTypes = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All types are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All types are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
-}
+};
 
 if (nconf.get('clean')) {
   deleteTypes()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing types...');
   importTypes()
+}
+module.exports = {
+  deleteTypes
 }
