@@ -1,9 +1,14 @@
-import {
-  client,
-  cartsService
-} from './services'
+import { client, cartsService } from './services'
+const nconf = require('nconf');
 
-const deleteAllCarts = () => {
+const deleteAllCarts = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: cartsService
       .build(),
@@ -27,8 +32,17 @@ const deleteAllCarts = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All carts are succesfully deleted')
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
 }
 
-deleteAllCarts()
+if (nconf.get('clean')) {
+  deleteAllCarts()
+}
+
+module.exports = {
+  deleteAllCarts
+}
