@@ -1,7 +1,14 @@
 import { client, channelsService } from './services'
-var nconf = require('nconf')
+const nconf = require('nconf');
 
-const deleteChannels = () => {
+const deleteChannels = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: channelsService
       .build(),
@@ -24,10 +31,13 @@ const deleteChannels = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All channels are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All channels are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
-}
+};
 
 const importChannels = () => {
   var channels = require(process.cwd() + '/data/channels.json')
@@ -46,5 +56,10 @@ const importChannels = () => {
 if (nconf.get('clean')) {
   deleteChannels()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing channels...');
   importChannels()
+}
+
+module.exports = {
+  deleteChannels
 }
