@@ -1,9 +1,17 @@
-import {
-  client,
-  inventoryService
-} from './services'
+import { client, inventoryService } from './services'
+require('dotenv').config();
+const nconf = require('nconf');
+const execSync = require('child_process').execSync;
+const path = require("path");
 
-const deleteInventory = () => {
+const deleteInventory = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: inventoryService
       .build(),
@@ -26,9 +34,27 @@ const deleteInventory = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All inventory entries are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All inventory entries are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
 }
 
-deleteInventory()
+const importInventory = (cmd) => {
+  console.log('Importing inventory...');
+  return path.join('node_modules', '.bin', cmd);
+}
+
+if (nconf.get('clean')) {
+  deleteInventory()
+} else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing inventory...');
+  execSync(`${importInventory('stock-import')} --logLevel debug --logSilent true --file data/inventory.csv --projectKey ${process.env.CONFIG_PROJECT_KEY} --clientId ${process.env.CONFIG_CLIENT_ID} --clientSecret ${process.env.CONFIG_CLIENT_SECRET} --sphereHost https://${process.env.CONFIG_API_URL} --sphereAuthHost https://${process.env.CONFIG_AUTH_URL}`);
+  execSync(`${importInventory('stock-import')} --logLevel debug --logSilent true --file data/inventory-stores.csv --projectKey ${process.env.CONFIG_PROJECT_KEY} --clientId ${process.env.CONFIG_CLIENT_ID} --clientSecret ${process.env.CONFIG_CLIENT_SECRET} --sphereHost https://${process.env.CONFIG_API_URL} --sphereAuthHost https://${process.env.CONFIG_AUTH_URL}`);
+}
+
+module.exports = {
+  deleteInventory
+}
