@@ -4,7 +4,7 @@ import { createAuthMiddlewareForClientCredentialsFlow } from '@commercetools/sdk
 import { createHttpMiddleware } from '@commercetools/sdk-middleware-http'
 import { createRequestBuilder, features } from '@commercetools/api-request-builder'
 require('es6-promise').polyfill()
-var fetch = require('isomorphic-fetch')
+const fetch = require('isomorphic-fetch')
 
 export const client = createClient({
   // The order of the middlewares is important !!!
