@@ -1,21 +1,20 @@
 import { client, shippingMethodService, taxService, zonesService } from './services'
-var nconf = require('nconf')
-
-var shippingMethods = require(process.cwd() + '/data/shipping-methods.json')
+const nconf = require('nconf')
+const shippingMethods = require(process.cwd() + '/data/shipping-methods.json')
 
 const getTaxCategories = () => {
   return client.execute({
     uri: taxService.build(),
     method: 'GET'
   }).catch(error => console.log(error))
-}
+};
 
 const getZones = () => {
   return client.execute({
     uri: zonesService.build(),
     method: 'GET'
   }).catch(error => console.log(error))
-}
+};
 
 const addShippingMethods = (taxCategoriesByName, zonesForName) => {
   shippingMethods.forEach(shippingMethod => {
@@ -33,9 +32,16 @@ const addShippingMethods = (taxCategoriesByName, zonesForName) => {
       body: shippingMethod
     }).catch(error => console.log(error.body.errors))
   })
-}
+};
 
-const deleteAllShippingMethods = () => {
+const deleteAllShippingMethods = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: shippingMethodService
       .build(),
@@ -58,12 +64,15 @@ const deleteAllShippingMethods = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All shipping methods are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All shipping methods are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
-}
+};
 
-async function run () {
+async function run (){
   try {
     let taxCategories = await getTaxCategories()
     const taxCategoriesByName = taxCategories.body.results.reduce((obj, taxCategory) => {
@@ -82,10 +91,15 @@ async function run () {
     console.log(e)
     throw e // let caller know the promise rejected with this reason
   }
-}
+};
 
 if (nconf.get('clean')) {
   deleteAllShippingMethods()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing shipping methods...');
   run()
+}
+
+module.exports = {
+  deleteAllShippingMethods
 }
