@@ -1,7 +1,14 @@
 import { client, taxService } from './services'
-var nconf = require('nconf')
+const nconf = require('nconf')
 
-const deleteTaxCategories = () => {
+const deleteTaxCategories = (cb) => {
+  let callerName;
+  try { throw new Error(); }
+  catch (e) { 
+    let re = /(\w+)@|at (\w+) \(/g, st = e.stack, m;
+    re.exec(st), m = re.exec(st);
+    callerName = m[1] || m[2];
+  };
   const request = {
     uri: taxService
       .build(),
@@ -24,10 +31,13 @@ const deleteTaxCategories = () => {
       )
     }
   ).then(() => {
-    console.log('\x1b[32m%s\x1b[0m', 'All tax categories are succesfully deleted')
+    console.log('\x1b[32m%s\x1b[0m', 'All tax categories are succesfully deleted');
+    if (callerName === 'clean'){
+      return cb();
+    }
   })
     .catch(err => console.log(err))
-}
+};
 
 const importTaxCategories = () => {
   var taxCategories = require(process.cwd() + '/data/tax-category.json')
@@ -46,5 +56,10 @@ const importTaxCategories = () => {
 if (nconf.get('clean')) {
   deleteTaxCategories()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing tax categories...');
   importTaxCategories()
+}
+
+module.exports = {
+  deleteTaxCategories
 }
